Metadata and Preparations
Added metadata about exam. Added ability to hash each question on the way into the store. Added ability to track incorrect questions. Added exam states to prepare for ability to pause or finish exam prematurely. Added several properties to the Exam Component in preparation to initialize exam with a timer, to show answer on incorrect response, to show the explanation, and to disable incorrect question tracking.